Refactor CLI imports and fix test imports
