+++ src/render_engine/__main__.py
@@ -1,6 +1,6 @@
 """Enables the use of `python -m render_engine` to run the CLI."""
 
-from .cli import app
+from .cli.cli import app
 
 if __name__ == "__main__":
     app()
