+++ src/render_engine/cli/__init__.py
@@ -1,0 +1,3 @@
+from .cli import app
+
+__main__ = ["app"]
